Move saving and loading to seperate file
