+++ BurntOut_Unity/Assets/Scripts/Editor/Narrative/NarrativeEditorWindow.cs
@@ -103,104 +103,30 @@ namespace Narrative.Inspector
 
         private void SaveScenarioBtn_Pressed(object sender, EventArgs e)
         {
-            // Get the folder to save the scenario in
-            string json = JsonUtility.ToJson(sceneNarrative.scenarios[tabs.Value - 1]);
-            var path = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
-            path += "\\BurntOut\\Narrative\\" + SceneManager.GetActiveScene().name + "\\Scenario";
-            Directory.CreateDirectory(path);
-
-            // Name.yyyy.MM.dd.letter.json
-            string fileName = sceneNarrative.scenarios[tabs.Value - 1].name + DateTime.Now.ToString("'.'yyyy'.'MM'.'dd");
-
-            // Alphabetical character based on number of similar files saved today
-            char lastLetter = 'a';
-            foreach (var file in Directory.GetFiles(path))
-            {
-                if (file.Contains(fileName) && file.EndsWith(".json"))
-                {
-                    string str = Path.GetFileName(file);
-                    str = str.Replace(fileName + '.', "");
-                    if (str[0] >= lastLetter)
-                        lastLetter = (char)(str[0] + 1);
-                }
-            }
-            fileName += "." + lastLetter + ".json";
-
-            path = EditorUtility.SaveFilePanel("Save Scenario", path, fileName, "json");
-
-            if (!String.IsNullOrEmpty(path))
-                File.WriteAllText(path, json);
+            NarrativeFileManager.SaveScenario(sceneNarrative.scenarios[tabs.Value - 1]);
         }
 
         private void LoadScenarioBtn_Pressed(object sender, EventArgs e)
         {
-            // Get the folder to load the scenario from
-            var path = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
-            path += "\\BurntOut\\Narrative\\" + SceneManager.GetActiveScene().name + "\\Scenario";
+            var scenario = NarrativeFileManager.LoadScenario(sceneNarrative);
+            if (scenario == null)
+                return;
 
-            path = EditorUtility.OpenFilePanel("Load Scenario", path, "json");
-
-            if (!String.IsNullOrEmpty(path))
-            {
-                string json = File.ReadAllText(path);
-
-                var scenario = JsonUtility.FromJson<Scenario>(json);
-                for (int i = 0; i < sceneNarrative.scenarios.Count; i++)
-                    if (i != tabs.Value - 1 && scenario.id == sceneNarrative.scenarios[i].id)
-                        scenario.ResetHash(sceneNarrative.scenarios.ToArray());
-
-                sceneNarrative.scenarios[tabs.Value - 1] = scenario;
-                scenarioEditors[tabs.Value - 1] = new ScenarioEditor(scenario);
-            }
-
-            OnEnable();
+            sceneNarrative.scenarios[tabs.Value - 1] = scenario;
+            scenarioEditors[tabs.Value - 1] = new ScenarioEditor(scenario);
         }
 
         private void SaveAllBtn_Pressed(object sender, EventArgs e)
         {
-            // Get the folder to save the scene in
-            string json = JsonUtility.ToJson(sceneNarrative);
-            var path = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
-            path += "\\BurntOut\\Narrative\\" + SceneManager.GetActiveScene().name;
-            Directory.CreateDirectory(path);
-
-            // Name.yyyy.MM.dd.letter.json
-            string fileName = SceneManager.GetActiveScene().name + DateTime.Now.ToString("'.'yyyy'.'MM'.'dd");
-
-            // Alphabetical character based on number of similar files saved today
-            char lastLetter = 'a';
-            foreach (var file in Directory.GetFiles(path))
-            {
-                if (file.Contains(fileName) && file.EndsWith(".json"))
-                {
-                    string str = Path.GetFileName(file);
-                    str = str.Replace(fileName + '.', "");
-                    if (str[0] >= lastLetter)
-                        lastLetter = (char)(str[0] + 1);
-                }
-            }
-            fileName += "." + lastLetter + ".json";
-
-            path = EditorUtility.SaveFilePanel("Save Scene", path, fileName, "json");
-
-            if (!String.IsNullOrEmpty(path))
-                File.WriteAllText(path, json);
+            NarrativeFileManager.SaveSceneNarrative(sceneNarrative);
         }
         private void LoadAllBtn_Pressed(object sender, EventArgs e)
         {
-            // Get the folder to load the scene from
-            var path = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
-            path += "\\BurntOut\\Narrative\\" + SceneManager.GetActiveScene().name;
+            var narrative = NarrativeFileManager.LoadSceneNarrative();
+            if (narrative == null)
+                return;
 
-            path = EditorUtility.OpenFilePanel("Load Scene", path, "json");
-
-            if (!String.IsNullOrEmpty(path))
-            {
-                string json = File.ReadAllText(path);
-
-                var narrative = JsonUtility.FromJson<SceneNarrative>(json);
-                sceneManager.sceneNarrative = narrative;
-            }
+            sceneManager.sceneNarrative = narrative;
             OnEnable();
         }
