+++ BurntOut_Unity/Assets/Scripts/Editor/Narrative/NarrativeFileManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 56044ee0e5dbfb848bf3b5758e988954
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
