+++ BurntOut_Unity/Assets/Scripts/Editor/OOEditor/Abstract/External.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: df8ae33324d501d4dbba567609097593
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
