+++ BurntOut_Unity/Assets/Scripts/Editor/OOEditor/Abstract/External/OOEditorWindow.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEditor;
+using UnityEngine;
+
+public class OOEditorWindow : EditorWindow {
+    
+}
