+++ BurntOut_Unity/Assets/Scripts/Editor/OOEditor/Abstract/External/OOEditorWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ca8ec12a5212ac4aaff6c5c627db585
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
